+++ CHANGELOG.md
@@ -37,6 +37,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+- [#297](https://github.com/babylonlabs-io/babylon/pull/297) Update relayer dependency
+
 ## v0.17.1
 
 ### Bug fixes