+++ CHANGELOG.md
@@ -42,6 +42,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - [#305](https://github.com/babylonlabs-io/babylon/pull/305) chore: add more error logs to `VerifyInclusionProofAndGetHeight`
 - [#304](https://github.com/babylonlabs-io/babylon/pull/304) Add highest voted height to finality provider
 
+### State Machine Breaking
+
+- [#310](https://github.com/babylonlabs-io/babylon/pull/310) implement adr-37 -
+making params valid for btc light client ranges
+
 ## v0.17.1
 
 ### Bug fixes